backend/requirements.txt
# PropertyPulse Python backend — FastAPI + Supabase + Stripe
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
supabase>=2.10.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
httpx>=0.27.0
stripe>=11.0.0
openai>=1.0.0
anthropic>=0.80.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
beautifulsoup4>=4.12.0
python-multipart>=0.0.9
